test(run_convobox): _working_watchdog coverage (C3) + ruff 0.16.2 drift fix (#143) (#273)
* test(run_convobox): direct coverage for _working_watchdog (issue #235, C3)

C3 flagged _working_watchdog as untested despite owning three real
responsibilities: approval-timeout auto-decline (the fail-closed
"silence must never be treated as consent" invariant), queued-
interjection delivery once the backend goes idle, and the TUI/web
status-computation priority ladder (paused > approval-waiting >
continue-waiting > speaking > working > capturing > listening). Its
constituent gate classes (ApprovalPromptGate, ContinuePromptGate,
QueuedInterjection, WorkingIndicator) already had direct tests
elsewhere -- none of them covered the watchdog's own orchestration of
them together.

16 new tests using a deterministic counted-sleep harness: monkeypatched
asyncio.sleep counts its own calls and raises CancelledError on the
(n+1)th, ending the watchdog's `while True` after exactly n body
iterations regardless of how many raw event-loop turns each iteration's
fake, non-suspending awaits actually take (a naive fixed-count
`asyncio.sleep(0)` loop in the test undercounts, since the watchdog's
own top-of-body sleep already consumes one turn before any tick body
runs -- caught by an earlier failing draft of this file, not asserted
from a docstring alone).

Verified: full suite green (1420 passed, +16), ruff clean, ensure_future
typing tightened for mypy (the one remaining mypy note on this file --
convobox.tui.state's missing py.typed marker -- is a standalone-check
artifact shared by every other test file that imports it; CI's mypy
invocation doesn't cover tests/, confirmed against test_run_convobox_quit.py).


* chore: lock ruff to the CI-resolved 0.16.2, fix new-rule violations (issue #143)

Local uv.lock was still pinned to ruff 0.15.21 while CI's dev-rig
workflows (which don't read this repo's lockfile, per issue #143's own
diagnosis) resolve whatever's newest within pyproject.toml's `>=0.15,<0.17`
bound -- 0.16.2 as of today. Two live incidents already hit this drift
(#236/#238 both failed CI-only on UP037). Re-resolving the lockfile to
0.16.2 and fixing what it flags locally closes that gap without touching
dev-rig (which isn't checked out on this machine) -- the actual
recommended fix (switch dev-rig's workflows to `uv sync`) still stands
per issue #143's own comment thread.

69 auto-fixed, all mechanical, zero behavior change: import sorting
(I001), stale `# noqa: E402` comments now unused because sorting moved
the import (RUF100), redundant quotes on forward-ref type annotations
that `from __future__ import annotations` already makes unnecessary
(UP037), `asyncio.TimeoutError` -> the builtin `TimeoutError` it's been
an alias of since Python 3.11 (UP041), a nested
`with pytest.raises(...): with X():` collapsed to one `with` statement
(SIM117, same runtime behavior). Reviewed every touched file's diff
directly, including the two safety-adjacent ones (run_convobox.py,
settings_tui.py) and every approval/barge-in/echo test file the sort
touched -- confirmed import-order and quote-removal only, no logic
changed.

25 new-rule findings remain (BLE001, ASYNC221, ISC004, PLW1510, RUF046,
RUF012, PYI034, DTZ005, RUF059) -- none auto-fixable, none in
safety-critical logic (scripts/voice_picker*.py, scripts/voice_tui.py,
scripts/audio_devices.py, scripts/ci_local.py, plus two false-positive
os.system("") flags in run_convobox.py's already-`# nosec`-annotated
VT-mode-enable calls); left for a follow-up pass rather than folding
judgment calls into this mechanical commit.

Verified: full suite green (1420 passed, matches pre-upgrade baseline),
ruff 0.16.2 clean except the 25 noted above, mypy unaffected.


* chore: resolve the remaining ruff 0.16.2 findings (issue #143 follow-up)

The 25 non-auto-fixable findings left over from the prior commit's
mechanical batch, triaged individually -- none in safety-critical logic
(safeword/approval/barge-in/VAD/STT), all in CLI tools, test fakes, or
already-deliberate patterns the new rule set doesn't recognize:

- BLE001 (6x, voice_picker.py/voice_picker_tui.py): CLI/TUI error
  boundaries that already carried an explanatory comment ("report and
  keep looping, not fatal") -- annotated `# noqa: BLE001` with that same
  reasoning, matching the `# noqa: BLE001 -- ...` style already
  established in audio_devices.py rather than narrowing the except
  clause and risking a real exception type getting missed in a CLI
  loop that must never crash on a bad download/audition.
- ASYNC221 (3x, run_convobox.py x2 + voice_tui.py): false positive on
  `os.system("")` -- each site already carries a `# nosec B605 B607`
  comment explaining it's a hardcoded empty-string VT-mode-enable idiom
  that spawns no process; added the matching `# noqa: ASYNC221`.
- ISC004 (4x, settings_tui.py x2 + voice_tui.py x2): deliberate
  multi-line f-string formatting the rule misreads as "forgot a comma"
  -- wrapped each in parens per ruff's own suggested fix, no string
  content changed.
- RUF012 (2x, test_audio.py): mutable class-attribute defaults on test
  fakes -- annotated `ClassVar`, matching their actual usage (shared
  across instances by design, not a per-instance default).
- RUF059 (3x): unused unpacked variables in tests (`reader2`, `adapter`,
  `value`) -- prefixed with `_`, including the one in
  test_orchestrator.py's hard_stop() coverage (a safety-critical test,
  though only the tuple-unpacking changed, not any assertion).
- PLW1510 (2x): `subprocess.run` without explicit `check` -- both sites
  already inspect `.returncode` manually right after, so `check=False`
  documents the existing intent rather than changing it.
- RUF046 (2x, audio_devices.py): `int(round(...))` -- `round()` with no
  ndigits already returns int; removed the redundant cast.
- PYI034 (2x): test fakes' `__enter__` returning their own class name
  instead of `Self` -- updated both, `Self` needs no extra dependency
  (requires-python is already >=3.12).
- DTZ005 (1x, acoustic_calibration.py): `datetime.now()` for an
  operator-browsed output-directory name -- deliberately local wall
  clock, not something that should switch to UTC; documented and
  suppressed rather than changed.

Verified: full suite green (1420 passed), `ruff check .` now fully
clean at 0.16.2, mypy unaffected (same 4 pre-existing yaml-stub notes
CI's types-PyYAML input already resolves).


* fix(tests): address code-quality bot review on PR #273

Two inline findings from github-code-quality[bot]:
- Module was imported both `import scripts.run_convobox as rc_mod` (to
  reach its `asyncio` attribute) and via the file's own top-level `from
  scripts.run_convobox import (...)`. Unnecessary -- that module's
  `asyncio` attribute IS this file's own `import asyncio` (same module
  object), so patching either reaches the same target. Removed the
  local import and monkeypatch `asyncio.sleep` directly, matching the
  pattern test_orchestrator.py's own _consume_events retry tests
  already use.
- "Statement has no effect" on the bare `await task` inside
  `pytest.raises(asyncio.CancelledError)` -- a false positive (the
  statement's effect IS the exception it raises, which the context
  manager catches), but added a one-line comment making that explicit
  rather than leaving a bot-flagged line unexplained.

Verified: full suite green (1423 passed), ruff clean.


* fix(tests): assign awaited task result to satisfy code-quality bot

github-code-quality flagged the bare `await task` in _run_ticks as a
statement with no effect on PR #273, blocking merge via required
conversation resolution. The await's side effect (raising
CancelledError) is intentional per the existing comment; assigning to
`_` keeps that behavior while giving the statement a visible effect.


---------

+++ scripts/audio_devices.py
@@ -250,7 +250,7 @@ def format_level(rms_db: float, peak_db: float, width: int = 30) -> str:
     """A text VU meter: RMS bar over a -60..0 dBFS range, plus a verdict (pure)."""
     floor_db = -60.0
     frac = max(0.0, min(1.0, (rms_db - floor_db) / (0.0 - floor_db)))
-    filled = int(round(frac * width))
+    filled = round(frac * width)
     bar = "#" * filled + "-" * (width - filled)
     return (
         f"[{bar}] rms {rms_db:6.1f} dBFS  peak {peak_db:6.1f} dBFS  "
@@ -567,7 +567,7 @@ def _resample_audio(audio: Any, src_rate: int, dst_rate: int) -> Any:
     audio = np.asarray(audio, dtype=np.float32)
     if src_rate == dst_rate or audio.size == 0:
         return audio
-    n_dst = int(round(len(audio) * dst_rate / src_rate))
+    n_dst = round(len(audio) * dst_rate / src_rate)
     if n_dst <= 0:
         return np.zeros(0, dtype=np.float32)
     src_x = np.linspace(0.0, 1.0, num=len(audio), endpoint=False)
